+++ src/test/kotlin/ecommerce/ProductMock.kt
@@ -1,6 +1,6 @@
 package ecommerce
 
-import ecommerce.product.Product
+import ecommerce.model.Product
 import io.restassured.RestAssured
 import io.restassured.http.ContentType
 import org.assertj.core.api.Assertions.assertThat
