+++ src/test/kotlin/ecommerce/repository/ProductRepositoryTest.kt
@@ -1,7 +1,7 @@
 package ecommerce.repository
 
-import ecommerce.product.Product
-import ecommerce.store.ProductStore
+import ecommerce.model.Product
+import ecommerce.repository.ProductStore
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
